Move kafka-watcher source code to exp
